requirements/dev.txt
fastapi==0.63.0
uvicorn==0.13.3
SQLAlchemy==1.3.23
alembic==1.5.4
psycopg2-binary==2.8.6
Pillow==8.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.2.0